# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qi local laboratory",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/collections/{collection_id}": {
      "get": {
        "summary": "View",
        "operationId": "view_api_collections__collection_id__get",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Run"
            }
          },
          {
            "name": "policy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Policy"
            }
          },
          {
            "name": "split",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Split"
            }
          },
          {
            "name": "disposition",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Disposition"
            }
          },
          {
            "name": "outcome",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Outcome"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "phase",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "",
                "opening",
                "middlegame",
                "endgame",
                "unknown"
              ],
              "type": "string",
              "default": "",
              "title": "Phase"
            }
          },
          {
            "name": "lens",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "shortfall",
                "long"
              ],
              "type": "string",
              "default": "all",
              "title": "Lens"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "newest",
                "longest",
                "shortfall"
              ],
              "type": "string",
              "default": "newest",
              "title": "Sort"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          },
          {
            "name": "attempts",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 40000,
              "default": "",
              "title": "Attempts"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionPage"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/collections/{collection_id}/quality": {
      "get": {
        "summary": "Quality",
        "operationId": "quality_api_collections__collection_id__quality_get",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionQuality"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollectionEntry": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          },
          "updated": {
            "type": "number",
            "title": "Updated"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "location",
          "bytes",
          "updated"
        ],
        "title": "CollectionEntry"
      },
      "CollectionPage": {
        "properties": {
          "collection": {
            "$ref": "#/components/schemas/CollectionEntry"
          },
          "as_of": {
            "type": "number",
            "title": "As Of"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/GeneratedRun"
            },
            "type": "array",
            "title": "Runs"
          },
          "overall": {
            "$ref": "#/components/schemas/CollectionStats"
          },
          "run_stats": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CollectionStats"
            },
            "type": "object",
            "title": "Run Stats"
          },
          "filtered": {
            "$ref": "#/components/schemas/CollectionStats"
          },
          "games": {
            "items": {
              "$ref": "#/components/schemas/GeneratedGame"
            },
            "type": "array",
            "title": "Games"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "collection",
          "as_of",
          "runs",
          "overall",
          "run_stats",
          "filtered",
          "games",
          "total",
          "offset",
          "limit"
        ],
        "title": "CollectionPage"
      },
      "CollectionQuality": {
        "properties": {
          "as_of": {
            "type": "number",
            "title": "As Of"
          },
          "selected_occurrences": {
            "type": "integer",
            "title": "Selected Occurrences"
          },
          "unique_inputs": {
            "type": "integer",
            "title": "Unique Inputs"
          },
          "cross_split_inputs": {
            "type": "integer",
            "title": "Cross Split Inputs"
          },
          "affected_train": {
            "type": "integer",
            "title": "Affected Train"
          },
          "affected_validation": {
            "type": "integer",
            "title": "Affected Validation"
          },
          "phases": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Phases"
          },
          "single_pv_budget_coverage": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Single Pv Budget Coverage"
          },
          "specs": {
            "items": {
              "$ref": "#/components/schemas/SpecCoverage"
            },
            "type": "array",
            "title": "Specs"
          },
          "overlap_examples": {
            "items": {
              "$ref": "#/components/schemas/OverlapExample"
            },
            "type": "array",
            "title": "Overlap Examples"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "selected_occurrences",
          "unique_inputs",
          "cross_split_inputs",
          "affected_train",
          "affected_validation",
          "phases",
          "single_pv_budget_coverage",
          "specs",
          "overlap_examples"
        ],
        "title": "CollectionQuality"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "GeneratedRun": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "identity": {
            "type": "string",
            "title": "Identity"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "planned": {
            "type": "integer",
            "title": "Planned"
          },
          "continued_from": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Continued From"
          },
          "created": {
            "type": "number",
            "title": "Created"
          }
        },
        "type": "object",
        "required": [
          "id",
          "identity",
          "name",
          "status",
          "planned",
          "continued_from",
          "created"
        ],
        "title": "GeneratedRun"
      },
...
      "SpecCoverage": {
        "properties": {
          "identity": {
            "type": "string",
            "title": "Identity"
          },
          "nodes": {
            "type": "integer",
            "title": "Nodes"
          },
          "depth": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Depth"
          },
          "multipv": {
            "type": "integer",
            "title": "Multipv"
          },
          "threads": {
            "type": "string",
            "title": "Threads"
          },
          "engine": {
            "type": "string",
            "title": "Engine"
          },
          "network": {
            "type": "string",
            "title": "Network"
          },
          "occurrences": {
            "type": "integer",
            "title": "Occurrences"
          }
        },
        "type": "object",
        "required": [
          "identity",
          "nodes",
          "depth",
          "multipv",
          "threads",
          "engine",
          "network",
          "occurrences"
        ],
        "title": "SpecCoverage"
      },
...
    }
  }
}